# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AIGovOps REPLAY API - cryptographically signed AI interaction receipts
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export interface LogoutSuccess {
  success: boolean;
}

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export type InteractionPolicyStatus =
  (typeof InteractionPolicyStatus)[keyof typeof InteractionPolicyStatus];

export const InteractionPolicyStatus = {
  pass: "pass",
  fail: "fail",
  pending: "pending",
} as const;

export interface Interaction {
  id: string;
  prompt: string;
  response: string;
  model: string;
  userId: string;
  tags: string[];
  promptHash: string;
  responseHash: string;
  prevHash?: string | null;
  chainHash: string;
  policyStatus: InteractionPolicyStatus;
  policyViolations: string[];
  createdAt: string;
  replayCount: number;
}

export interface InteractionList {
  items: Interaction[];
  total: number;
  limit: number;
  offset: number;
}

export interface CreateInteractionBody {
  prompt: string;
  response: string;
  model: string;
  tags?: string[];
}

export interface VerificationResult {
  id: string;
  valid: boolean;
  promptHashMatch: boolean;
  responseHashMatch: boolean;
  chainIntact: boolean;
  details: string;
  checkedAt: string;
}

export interface ReplayResult {
  originalId: string;
  originalResponse: string;
  replayedResponse: string;
  outputDiff: string;
  semanticMatch: boolean;
  replayedAt: string;
  newReceiptId: string;
}

export type PolicySeverity =
  (typeof PolicySeverity)[keyof typeof PolicySeverity];

export const PolicySeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Policy {
  id: string;
  name: string;
  description: string;
  rule: string;
  severity: PolicySeverity;
  enabled: boolean;
  violationCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface PolicyList {
  items: Policy[];
  total: number;
}

export type CreatePolicyBodySeverity =
  (typeof CreatePolicyBodySeverity)[keyof typeof CreatePolicyBodySeverity];

export const CreatePolicyBodySeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreatePolicyBody {
  name: string;
  description: string;
  rule: string;
  severity: CreatePolicyBodySeverity;
  enabled?: boolean;
}

export type UpdatePolicyBodySeverity =
  (typeof UpdatePolicyBodySeverity)[keyof typeof UpdatePolicyBodySeverity];

export const UpdatePolicyBodySeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface UpdatePolicyBody {
  name?: string;
  description?: string;
  rule?: string;
  severity?: UpdatePolicyBodySeverity;
  enabled?: boolean;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  created: "created",
  replayed: "replayed",
  verified: "verified",
  policy_check: "policy_check",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  interactionId: string;
  summary: string;
  createdAt: string;
}

export interface Stats {
  totalInteractions: number;
  verifiedCount: number;
  policyPassCount: number;
  policyFailCount: number;
  replayCount: number;
  modelsUsed: string[];
  recentActivity: ActivityItem[];
  /** Total number of receipts in the chain (full count, not a 100-entry window) */
  chainLength: number;
  /** True only when the full chain has no broken links and no forks (checked over all receipts) */
  chainIntact: boolean;
}

export type ChainSummaryEntriesItem = {
  id: string;
  chainHash: string;
  prevHash?: string | null;
  createdAt: string;
};

export interface ChainSummary {
  /** Total number of receipts in the chain (full count, not a window) */
  length: number;
  headHash: string;
  tailHash?: string | null;
  /** True only when the full chain has no broken links and no forks (checked over all receipts, not a window) */
  intact: boolean;
  /** Number of prevHash values claimed by more than one receipt (indicates concurrent write races) */
  forkCount: number;
  /** Number of receipts whose prevHash does not match any existing chainHash */
  brokenLinkCount: number;
  /** Number of receipts with null prevHash; a valid non-empty chain has exactly one */
  genesisCount: number;
  /** Most recent receipts (up to 100) for display. Integrity is verified over the full chain. */
  entries: ChainSummaryEntriesItem[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListInteractionsParams = {
  limit?: number;
  offset?: number;
  model?: string;
  policyStatus?: ListInteractionsPolicyStatus;
};

export type ListInteractionsPolicyStatus =
  (typeof ListInteractionsPolicyStatus)[keyof typeof ListInteractionsPolicyStatus];

export const ListInteractionsPolicyStatus = {
  pass: "pass",
  fail: "fail",
  pending: "pending",
} as const;