# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIGovOps REPLAY API - cryptographically signed AI interaction receipts
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interactions:
    get:
      operationId: listInteractions
      tags: [interactions]
      summary: List all interaction receipts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: model
          in: query
          schema:
            type: string
        - name: policyStatus
          in: query
          schema:
            type: string
            enum: [pass, fail, pending]
      responses:
        "200":
          description: List of interaction receipts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractionList"
    post:
      operationId: createInteraction
      tags: [interactions]
      summary: Create a new AI interaction receipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInteractionBody"
      responses:
        "201":
          description: Created interaction receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"
  /interactions/{id}:
    get:
      operationId: getInteraction
      tags: [interactions]
      summary: Get a single interaction receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Interaction receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"
        "404":
          description: Not found
  /interactions/{id}/verify:
    get:
      operationId: verifyInteraction
      tags: [interactions]
      summary: Verify the cryptographic chain for a receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
  /interactions/{id}/replay:
    post:
      operationId: replayInteraction
      tags: [interactions]
      summary: Replay an interaction and compare outputs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Replay result with diff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayResult"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token passed as a Bearer token in the Authorization header (used by mobile clients)
...
  schemas:
...
    Interaction:
      type: object
      properties:
        id:
          type: string
        prompt:
          type: string
        response:
          type: string
        model:
          type: string
        userId:
          type: string
        tags:
          type: array
          items:
            type: string
        promptHash:
          type: string
        responseHash:
          type: string
        prevHash:
          type: string
          nullable: true
        chainHash:
          type: string
        policyStatus:
          type: string
          enum: [pass, fail, pending]
        policyViolations:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        replayCount:
          type: integer
      required:
        - id
        - prompt
        - response
        - model
        - userId
        - tags
        - promptHash
        - responseHash
        - chainHash
        - policyStatus
        - policyViolations
        - createdAt
        - replayCount
    InteractionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Interaction"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
    CreateInteractionBody:
      type: object
      properties:
        prompt:
          type: string
        response:
          type: string
        model:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - prompt
        - response
        - model
    VerificationResult:
      type: object
      properties:
        id:
          type: string
        valid:
          type: boolean
        promptHashMatch:
          type: boolean
        responseHashMatch:
          type: boolean
        chainIntact:
          type: boolean
        details:
          type: string
        checkedAt:
          type: string
          format: date-time
      required:
        - id
        - valid
        - promptHashMatch
        - responseHashMatch
        - chainIntact
        - details
        - checkedAt
    ReplayResult:
      type: object
      properties:
        originalId:
          type: string
        originalResponse:
          type: string
        replayedResponse:
          type: string
        outputDiff:
          type: string
        semanticMatch:
          type: boolean
        replayedAt:
          type: string
          format: date-time
        newReceiptId:
          type: string
      required:
        - originalId
        - originalResponse
        - replayedResponse
        - outputDiff
        - semanticMatch
        - replayedAt
        - newReceiptId
...